refactor: decompose statement generation into modules
Break down the massive `statement.rs` file into logical submodules,
greatly improving code organization and maintainability.

Changes:
- New Directory Structure `llvm_temporary/statement/`:
  - `mod.rs`: Main entry point for `generate_statement_ir`.
  - `variable.rs`: Variable declaration and initialization (`ASTNode::Variable`).
  - `assign.rs`: Assignment (`ASTNode::Assign`).
  - `control.rs`: Control flow (`If`, `While`, `Break`, `Continue`, `Return`).
  - `io.rs`: Output/Input (`Print`, `Println`, `Input`).
  - `asm.rs`: Inline assembly blocks (`AsmBlock`).
  - `expr_stmt.rs`: Expression statements (`Expression`).
- Migration:
  - Moved logic from the single `generate_statement_ir` function into
    specialized functions within each module (e.g., `gen_variable_ir`,
    `gen_if_ir`).
  - Extracted helper functions like `truthy_to_i1` and `make_global_cstr`
    to appropriate modules.
- Functionality:
  - Preserved all existing logic.
  - No behavioral changes, purely structural refactoring.

This prepares the statement generation logic for future extensions and
makes the codebase easier to navigate.
